fix: resolve test isolation issues in split workspace test files
- Add missing fixtures to conftest.py (mock_project_dir, mock_worktree_path,
  workspace_spec_dir, with_spec_branch, with_conflicting_branches)
- Add module isolation fixture to test_cli_workspace_utils.py to restore
  workspace_commands module state after sys.modules manipulation tests
- Update tests to use workspace_spec_dir instead of spec_dir where needed
- Remove duplicate fixture definitions that were causing conflicts
