+++ tests/conftest.py
@@ -264,6 +264,154 @@ def spec_dir(temp_dir: Path) -> Path:
 
 
 # =============================================================================
+# WORKSPACE COMMAND TEST FIXTURES
+# =============================================================================
+
+# Constants for workspace tests
+TEST_SPEC_NAME = "001-test-spec"
+TEST_SPEC_BRANCH = f"auto-claude/{TEST_SPEC_NAME}"
+
+
+@pytest.fixture
+def mock_project_dir(temp_git_repo: Path) -> Path:
+    """Create a mock project directory with git repo."""
+    return temp_git_repo
+
+
+@pytest.fixture
+def mock_worktree_path(temp_git_repo: Path) -> Path:
+    """Create a mock worktree path."""
+    worktree_path = temp_git_repo / ".worktrees" / TEST_SPEC_NAME
+    worktree_path.mkdir(parents=True, exist_ok=True)
+    return worktree_path
+
+
+@pytest.fixture
+def workspace_spec_dir(temp_git_repo: Path) -> Path:
+    """Create a spec directory inside .auto-claude/specs/ for workspace tests."""
+    spec_dir = temp_git_repo / ".auto-claude" / "specs" / TEST_SPEC_NAME
+    spec_dir.mkdir(parents=True, exist_ok=True)
+    return spec_dir
+
+
+@pytest.fixture
+def with_spec_branch(temp_git_repo: Path) -> Generator[Path, None, None]:
+    """Create a temp git repo with a spec branch."""
+    # Create initial commit on main
+    (temp_git_repo / "README.md").write_text("# Test Repo")
+    subprocess.run(
+        ["git", "add", "README.md"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+    subprocess.run(
+        ["git", "commit", "-m", "Initial commit"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+
+    # Create spec branch
+    subprocess.run(
+        ["git", "checkout", "-b", TEST_SPEC_BRANCH],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+
+    # Add a change on spec branch
+    (temp_git_repo / "test.txt").write_text("test content")
+    subprocess.run(
+        ["git", "add", "test.txt"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+    subprocess.run(
+        ["git", "commit", "-m", "Test commit"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+
+    # Go back to main
+    subprocess.run(
+        ["git", "checkout", "main"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+
+    yield temp_git_repo
+
+
+@pytest.fixture
+def with_conflicting_branches(temp_git_repo: Path) -> Generator[Path, None, None]:
+    """Create temp git repo with conflicting branches for merge testing."""
+    # Create initial commit
+    (temp_git_repo / "README.md").write_text("# Test Repo")
+    subprocess.run(
+        ["git", "add", "README.md"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+    subprocess.run(
+        ["git", "commit", "-m", "Initial commit"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+
+    # Create spec branch
+    subprocess.run(
+        ["git", "checkout", "-b", TEST_SPEC_BRANCH],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+
+    # Add a file on spec branch
+    (temp_git_repo / "conflict.txt").write_text("spec branch content")
+    subprocess.run(
+        ["git", "add", "conflict.txt"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+    subprocess.run(
+        ["git", "commit", "-m", "Spec change"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+
+    # Go back to main and make conflicting change
+    subprocess.run(
+        ["git", "checkout", "main"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+    (temp_git_repo / "conflict.txt").write_text("main branch content")
+    subprocess.run(
+        ["git", "add", "conflict.txt"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+    subprocess.run(
+        ["git", "commit", "-m", "Main change"],
+        cwd=temp_git_repo,
+        capture_output=True,
+        check=True,
+    )
+
+    yield temp_git_repo
+
+
+# =============================================================================
 # REVIEW FIXTURES - Import from review_fixtures.py
 # =============================================================================
